+++ apps/nextjs/src/app/aila/resources/resources-contents.tsx
@@ -57,7 +57,8 @@ const ResourcesContentsInner: FC<AdditionalMaterialsUserProps> = () => {
   const { resetFormState } = useResourcesActions();
   const { setDialogWindow } = useDialog();
 
-  const { handleSubmitLessonPlan, handleSubmit } = useStepSubmitLogic();
+  const { handleSubmitLessonPlan, handleSubmit, handleCreateSession } =
+    useStepSubmitLogic();
 
   useEffect(() => {
     resetFormState();
@@ -110,7 +111,7 @@ const ResourcesContentsInner: FC<AdditionalMaterialsUserProps> = () => {
   };
 
   const stepComponents = {
-    0: <StepOne />,
+    0: <StepOne handleCreateSession={handleCreateSession} />,
     1: <StepTwo handleSubmitLessonPlan={handleSubmitLessonPlan} />,
     2: <StepThree handleSubmit={handleSubmit} />,
     3: <StepFour />,