+++ apps/nextjs/src/instrumentation/tracer.ts
@@ -1,7 +1,0 @@
-import { initializeTracer } from "@oakai/core/src/tracing/baseTracing";
-
-export function initTracer() {
-  initializeTracer({});
-}
-
-initTracer();