+++ CHANGE_LOG.md
@@ -1,3 +1,47 @@
+# [1.41.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.40.1...v1.41.0) (2025-06-12)
+
+
+### Bug Fixes
+
+* continue button case inconsistency and regenerate action bugs (AI-1261) ([#704](https://github.com/oaknational/oak-ai-lesson-assistant/issues/704)) ([f9a843f](https://github.com/oaknational/oak-ai-lesson-assistant/commit/f9a843ff5cff70d150696799c0e9c2a5f005f9ca))
+* remove helicone threat detection from moderation calls ([#683](https://github.com/oaknational/oak-ai-lesson-assistant/issues/683)) ([fa37fb9](https://github.com/oaknational/oak-ai-lesson-assistant/commit/fa37fb984cbba22d91f11553a4115623a1f971d1))
+
+
+### Features
+
+* improvements to glossary prompt and download file names AI-1264 ([#705](https://github.com/oaknational/oak-ai-lesson-assistant/issues/705)) ([edde171](https://github.com/oaknational/oak-ai-lesson-assistant/commit/edde171939cbf02dff5109abf6fb2b7188dfc786))
+* starter and exit quiz in slides ([#643](https://github.com/oaknational/oak-ai-lesson-assistant/issues/643)) ([b6a45fc](https://github.com/oaknational/oak-ai-lesson-assistant/commit/b6a45fc11b1f859db861b1ec6ad476501c4bf1a1))
+* title, subject, key-stage, topic agents, and error handling [AI-1217] ([#684](https://github.com/oaknational/oak-ai-lesson-assistant/issues/684)) ([e802922](https://github.com/oaknational/oak-ai-lesson-assistant/commit/e802922e346eb1716092075c5f6638c8cc7c8be5))
+* use fixed window rate limiting for demo users ([#699](https://github.com/oaknational/oak-ai-lesson-assistant/issues/699)) ([7c6cbc2](https://github.com/oaknational/oak-ai-lesson-assistant/commit/7c6cbc2ffd83ed526ef36fad3770366577a18e4f))
+
+
+### Performance Improvements
+
+* add Sentry Tracing for Performance Monitoring ([#708](https://github.com/oaknational/oak-ai-lesson-assistant/issues/708)) ([d007aac](https://github.com/oaknational/oak-ai-lesson-assistant/commit/d007aac3af01722b50224620851019270d89f465))
+
+## [1.40.1](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.40.0...v1.40.1) (2025-06-05)
+
+
+### Bug Fixes
+
+* force dynamic rendering for cron job routes ([#700](https://github.com/oaknational/oak-ai-lesson-assistant/issues/700)) ([79e280c](https://github.com/oaknational/oak-ai-lesson-assistant/commit/79e280ca04b4b0d8552734340baf52d07b6a99c1))
+
+# [1.40.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.39.0...v1.40.0) (2025-06-05)
+
+
+### Bug Fixes
+
+* add admin user button to slack alert when user banned ([#691](https://github.com/oaknational/oak-ai-lesson-assistant/issues/691)) ([d4ab295](https://github.com/oaknational/oak-ai-lesson-assistant/commit/d4ab29534dc3555329ceb42196dd9251cd14c820))
+
+
+### Features
+
+* add scheduled clerk user cleanup cron job - AI-1245 ([#693](https://github.com/oaknational/oak-ai-lesson-assistant/issues/693)) ([cc8b7d9](https://github.com/oaknational/oak-ai-lesson-assistant/commit/cc8b7d90fc9a25983164339fa6bceddce7496d3a))
+* agentic maths quiz [AI-1189] [AI-1152] ([#677](https://github.com/oaknational/oak-ai-lesson-assistant/issues/677)) ([3b7725c](https://github.com/oaknational/oak-ai-lesson-assistant/commit/3b7725ced2453c96f1d40701801e4b85915075b4))
+* am mobile implementation ([#696](https://github.com/oaknational/oak-ai-lesson-assistant/issues/696)) ([c87981f](https://github.com/oaknational/oak-ai-lesson-assistant/commit/c87981f7ac730e4e36e81148630b56d456abe16a))
+* change the steps ([#685](https://github.com/oaknational/oak-ai-lesson-assistant/issues/685)) ([e48d5f7](https://github.com/oaknational/oak-ai-lesson-assistant/commit/e48d5f768908181430953b64ac78eea9ad4b6024))
+* update modify hover states copy - AI-1246 ([#692](https://github.com/oaknational/oak-ai-lesson-assistant/issues/692)) ([a3e3424](https://github.com/oaknational/oak-ai-lesson-assistant/commit/a3e342457d4b3edd5dea8079f305f1b4d9ef15f3))
+
 # [1.39.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.38.0...v1.39.0) (2025-05-29)